requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
google-generativeai>=0.3.0